Migratory Birds
We have been asked to help study the population of birds migrating across the continent. Each type of bird we are interested in will be identified by an integer value. Each time a particular kind of bird is spotted, its id number will be added to our array of sightings. We would like to be able to find out which type of bird is most common given a list of sightings. Our task is to print the type number of that bird and if two or more types of birds are equally common, choose the type with the smallest ID number.